added the correct version

+++ handlers/rdiff_handler.c
@@ -19,7 +19,7 @@
 #if defined(__FreeBSD__)
 #include <sys/param.h>
 #endif
-#include "swupdate_image.h"
+#include "swupdate.h"
 #include "handler.h"
 #include "util.h"
 
@@ -173,11 +173,11 @@ static inline void rdiff_stats(const char* msg, struct rdiff_t *rdiff_state, rs_
 		case RS_RUNNING: strresult = (char*)"RUNNING"; break;
 		default: break;
 	}
-	TRACE("%s avail_in=%zu avail_out=%zu result=%s",
+	TRACE("%s avail_in=%ld avail_out=%ld result=%s",
 		  msg, buffers->avail_in, buffers->avail_out, strresult);
 }
 
-static int apply_rdiff_chunk_cb(void *out, const void *buf, size_t len)
+static int apply_rdiff_chunk_cb(void *out, const void *buf, unsigned int len)
 {
 	struct rdiff_t *rdiff_state = (struct rdiff_t *)out;
 	rs_buffers_t *buffers = &rdiff_state->buffers;
